requirements.txt
Django==5.0.1
Pillow==10.2.0
reportlab==4.0.7
tensorflow
torch
torchvision
requests
numpy==1.26.3
pandas==2.1.4
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
scikit-learn==1.3.2
supabase==2.3.4
gotrue==2.5.0
python-dotenv==1.0.0


